Adding jupyter notebook

+++ violajones/HaarLikeFeature.py
@@ -1,4 +1,4 @@
-import IntegralImage as ii
+import violajones.IntegralImage as ii
 
 
 def enum(**enums):
@@ -48,32 +48,32 @@ class HaarLikeFeature(object):
         """
         score = 0
         if self.type == FeatureType.TWO_VERTICAL:
-            first = ii.sum_region(int_img, self.top_left, (self.top_left[0] + self.width, self.top_left[1] + self.height / 2))
-            second = ii.sum_region(int_img, (self.top_left[0], self.top_left[1] + self.height / 2), self.bottom_right)
+            first = ii.sum_region(int_img, self.top_left, (self.top_left[0] + self.width, int(self.top_left[1] + self.height / 2)))
+            second = ii.sum_region(int_img, (self.top_left[0], int(self.top_left[1] + self.height / 2)), self.bottom_right)
             score = first - second
         elif self.type == FeatureType.TWO_HORIZONTAL:
-            first = ii.sum_region(int_img, self.top_left, (self.top_left[0] + self.width / 2, self.top_left[1] + self.height))
-            second = ii.sum_region(int_img, (self.top_left[0] + self.width / 2, self.top_left[1]), self.bottom_right)
+            first = ii.sum_region(int_img, self.top_left, (int(self.top_left[0] + self.width / 2), self.top_left[1] + self.height))
+            second = ii.sum_region(int_img, (int(self.top_left[0] + self.width / 2), self.top_left[1]), self.bottom_right)
             score = first - second
         elif self.type == FeatureType.THREE_HORIZONTAL:
-            first = ii.sum_region(int_img, self.top_left, (self.top_left[0] + self.width / 3, self.top_left[1] + self.height))
-            second = ii.sum_region(int_img, (self.top_left[0] + self.width / 3, self.top_left[1]), (self.top_left[0] + 2 * self.width / 3, self.top_left[1] + self.height))
-            third = ii.sum_region(int_img, (self.top_left[0] + 2 * self.width / 3, self.top_left[1]), self.bottom_right)
+            first = ii.sum_region(int_img, self.top_left, (int(self.top_left[0] + self.width / 3), self.top_left[1] + self.height))
+            second = ii.sum_region(int_img, (int(self.top_left[0] + self.width / 3), self.top_left[1]), (int(self.top_left[0] + 2 * self.width / 3), self.top_left[1] + self.height))
+            third = ii.sum_region(int_img, (int(self.top_left[0] + 2 * self.width / 3), self.top_left[1]), self.bottom_right)
             score = first - second + third
         elif self.type == FeatureType.THREE_VERTICAL:
-            first = ii.sum_region(int_img, self.top_left, (self.bottom_right[0], self.top_left[1] + self.height / 3))
-            second = ii.sum_region(int_img, (self.top_left[0], self.top_left[1] + self.height / 3), (self.bottom_right[0], self.top_left[1] + 2 * self.height / 3))
-            third = ii.sum_region(int_img, (self.top_left[0], self.top_left[1] + 2 * self.height / 3), self.bottom_right)
+            first = ii.sum_region(int_img, self.top_left, (self.bottom_right[0], int(self.top_left[1] + self.height / 3)))
+            second = ii.sum_region(int_img, (self.top_left[0], int(self.top_left[1] + self.height / 3)), (self.bottom_right[0], int(self.top_left[1] + 2 * self.height / 3)))
+            third = ii.sum_region(int_img, (self.top_left[0], int(self.top_left[1] + 2 * self.height / 3)), self.bottom_right)
             score = first - second + third
         elif self.type == FeatureType.FOUR:
             # top left area
-            first = ii.sum_region(int_img, self.top_left, (self.top_left[0] + self.width / 2, self.top_left[1] + self.height / 2))
+            first = ii.sum_region(int_img, self.top_left, (int(self.top_left[0] + self.width / 2), int(self.top_left[1] + self.height / 2)))
             # top right area
-            second = ii.sum_region(int_img, (self.top_left[0] + self.width / 2, self.top_left[1]), (self.bottom_right[0], self.top_left[1] + self.height / 2))
+            second = ii.sum_region(int_img, (int(self.top_left[0] + self.width / 2), self.top_left[1]), (self.bottom_right[0], int(self.top_left[1] + self.height / 2)))
             # bottom left area
-            third = ii.sum_region(int_img, (self.top_left[0], self.top_left[1] + self.height / 2), (self.top_left[0] + self.width / 2, self.bottom_right[1]))
+            third = ii.sum_region(int_img, (self.top_left[0], int(self.top_left[1] + self.height / 2)), (int(self.top_left[0] + self.width / 2), self.bottom_right[1]))
             # bottom right area
-            fourth = ii.sum_region(int_img, (self.top_left[0] + self.width / 2, self.top_left[1] + self.height / 2), self.bottom_right)
+            fourth = ii.sum_region(int_img, (int(self.top_left[0] + self.width / 2), int(self.top_left[1] + self.height / 2)), self.bottom_right)
             score = first - second - third + fourth
         return score
